cartoon_diary/apps/accounts/serializers.py
"""Serializers for account operations."""

from __future__ import annotations

from django.contrib.auth import authenticate, get_user_model
from rest_framework import serializers


User = get_user_model()


class SignupSerializer(serializers.ModelSerializer):
    password = serializers.CharField(write_only=True)

    class Meta:
        model = User
        fields = ("id", "email", "username", "password")

    def create(self, validated_data):
        username = validated_data.get("username") or validated_data["email"].split("@")[0]
        return User.objects.create_user(
            email=validated_data["email"],
            username=username,
            password=validated_data["password"],
        )


class LoginSerializer(serializers.Serializer):
    email = serializers.EmailField()
    password = serializers.CharField(write_only=True)

    def validate(self, attrs):
        user = authenticate(username=attrs["email"], password=attrs["password"])
        if not user:
            raise serializers.ValidationError("Invalid credentials")
        attrs["user"] = user
        return attrs
