cartoon_diary/apps/profiles/admin.py
"""Admin bindings for profiles."""

from django.contrib import admin

from .models import Profile


@admin.register(Profile)
class ProfileAdmin(admin.ModelAdmin):
    list_display = ("user", "nickname", "timezone", "created_at")

