cartoon_diary/requirements/dev.txt
-r base.txt

pytest>=8.1,<9.0
pytest-django>=4.8,<5.0
pytest-cov>=4.1,<5.0
factory-boy>=3.3,<4.0
django-debug-toolbar>=4.3,<5.0
